refactor + added some todos
