+++ src/Chess/Chess/GameAnalyzer.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Chess
+{
+    class GameAnalyzer
+    {
+        // TODO: validate moves here, e.g. mate, check mate... most likely rules here
+    }
+}
